Fix -Wconversion errors in single complex TESTING/EIG sources during Index-64 builds

+++ TESTING/EIG/cdrvbd.f
@@ -1140,8 +1140,8 @@
                   IL = 1
                   IU = MAX( 1, MNMIN )
                ELSE
-                  IL = 1 + INT( ( MNMIN-1 )*SLARND( 1, ISEED2 ) )
-                  IU = 1 + INT( ( MNMIN-1 )*SLARND( 1, ISEED2 ) )
+                  IL = 1 + INT( REAL( MNMIN-1 )*SLARND( 1, ISEED2 ) )
+                  IU = 1 + INT( REAL( MNMIN-1 )*SLARND( 1, ISEED2 ) )
                   IF( IU.LT.IL ) THEN
                      ITEMP = IU
                      IU = IL
